requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
python-dotenv==1.1.0
google-adk==0.5.0
pdfplumber==0.11.6